requirements.txt
transformers
torch
torchvision
numpy 
pillow
tqdm
einops
huggingface_hub
python-dotenv
sentencepiece